requirements.txt
Flask==3.0.2
gunicorn==21.2.0
waitress==2.1.2
python-dotenv==1.0.1
psycopg2==2.9.6
Flask-SQLAlchemy==3.0.0
yoyo-migrations==7.0.0
